Front: Move User Atoms to State folder
