+++ front/src/components/actions/auth.tsx
@@ -16,8 +16,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { accessTokenAtom } from "../../contexts/user";
 import { store } from "../../state/store";
+import { accessTokenAtom } from "../../state/user";
 import { LogoutIcon } from "../icons";
 import type Action from "./action";
