+++ front/src/components/actions/update-illustration.tsx
@@ -28,8 +28,8 @@ import { useTranslation } from "react-i18next";
 import { useMutation } from "react-query";
 import API from "../../api/api";
 import type { QueryClient } from "../../api/use-query";
-import { userAtom } from "../../contexts/user";
 import { store } from "../../state/store";
+import { userAtom } from "../../state/user";
 import { UpdateIllustrationIcon } from "../icons";
 import type Action from "./action";
