+++ front/src/components/authentication/authentication-form.tsx
@@ -23,7 +23,7 @@ import { useState } from "react";
 import { toast } from "react-hot-toast";
 import { useTranslation } from "react-i18next";
 import API from "../../api/api";
-import { accessTokenAtom } from "../../contexts/user";
+import { accessTokenAtom } from "../../state/user";
 
 /**
  * Authentication form
