+++ front/src/components/contextual-menu/track-contextual-menu.tsx
@@ -23,8 +23,8 @@ import { useTranslation } from "react-i18next";
 import { useMutation } from "react-query";
 import API from "../../api/api";
 import { useQueryClient } from "../../api/use-query";
-import { userAtom } from "../../contexts/user";
 import type { TrackWithRelations } from "../../models/track";
+import { userAtom } from "../../state/user";
 import type Action from "../actions/action";
 import { DownloadAction } from "../actions/download";
 import { GoToReleaseAction, GoToSongLyricsAction } from "../actions/link";
